+++ app/src/main/java/io/github/currency/companion/CurrencyApplication.kt
@@ -6,7 +6,7 @@ import io.github.currency.companion.data.*
 
 class CurrencyApplication : Application() {
     val repository by lazy {
-        CurrencyRepository(Room.databaseBuilder(this, CurrencyDatabase::class.java, "currency.db").addMigrations(CurrencyDatabase.MIGRATION_1_2).build())
+        CurrencyRepository(Room.databaseBuilder(this, CurrencyDatabase::class.java, "currency.db").addMigrations(CurrencyDatabase.MIGRATION_1_2, CurrencyDatabase.MIGRATION_2_3).build())
     }
     val gateway by lazy { SttGateway(this) }
 }